third_party/blink/web_tests/external/wpt/service-workers/service-worker/ServiceWorkerGlobalScope/message-event-ports.https-expected.txt

This is a testharness.js-based test.
[FAIL] Verify MessageEvent.ports getter returns the same object
  assert_equals: empty array expected "same ports array" but got "different ports array"
Harness: the test ran to completion.
